+++ src/Models/Configuration/Rules.php
@@ -10,4 +10,9 @@ interface Rules
      */
     public const NESTING_DEPTH = 'nestingDepth';
 
+    /**
+     *
+     */
+    public const KEY_LENGTH = 'keyLength';
+
 }